new changes

+++ src/test/java/StatsServiceTest.java
@@ -9,9 +9,9 @@ public class StatsServiceTest {
     @Test
     void calculateSum() {
         StatsService service = new StatsService();
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        double[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         long expected = 8 + 15 + 13 + 15 + 17 + 20 + 19 + 20 + 7 + 14 + 14 + 18;
-        long actual = service.calculateSumSales(sales);
+        double actual = service.calculateSumSales(sales);
         assertEquals(expected, actual);
     }
 
@@ -45,7 +45,7 @@ public class StatsServiceTest {
     @Test
     void findBelowAverage() {
         StatsService service = new StatsService();
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        double[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         long expected = 5;
         long actual = service.belowAverage(sales);
         assertEquals(expected, actual);
@@ -54,7 +54,7 @@ public class StatsServiceTest {
     @Test
     void findAbowAverage() {
         StatsService service = new StatsService();
-        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        double[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         long expected = 5;
         long actual = service.abowAverage(sales);
         assertEquals(expected, actual);
